test: adds RegressionConfirmationsRedesigned tag, split tests to improve test execution time consistency
